Configs: Update config files.
This updates the bundled config, and example config, to use wildcard table names. This makes sure that anyone using a default setup accounts for unknown table prefixes, and profiles the use in the example as well.

As a bonus, it means our tests now run against wildcard tables names for added accuracy!